Renamed firebase_ml to firebase_ml_custom
