+++ packages/firebase_ml/android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'firebase_ml'
+rootProject.name = 'firebase_ml_custom'
