+++ packages/firebase_ml/android/src/main/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="io.flutter.plugins.firebaseml">
+  package="io.flutter.plugins.firebasemlcustom">
     <application>
         <service android:name="com.google.firebase.components.ComponentDiscoveryService">
-            <meta-data android:name="com.google.firebase.components:io.flutter.plugins.firebaseml.FlutterFirebaseAppRegistrar"
+            <meta-data android:name="com.google.firebase.components:io.flutter.plugins.firebasemlcustom.FlutterFirebaseAppRegistrar"
                 android:value="com.google.firebase.components.ComponentRegistrar" />
         </service>
     </application>
