+++ packages/firebase_ml/android/src/main/java/io/flutter/plugins/firebasemlcustom/ModelManager.java
@@ -1,4 +1,4 @@
-package io.flutter.plugins.firebaseml;
+package io.flutter.plugins.firebasemlcustom;
 
 import android.os.Build;
 import androidx.annotation.NonNull;
