+++ packages/firebase_ml/android/user-agent.gradle
@@ -2,7 +2,7 @@ import java.util.regex.Matcher
 import java.util.regex.Pattern
 
 String libraryVersionName = "UNKNOWN"
-String libraryName = "flutter-fire-ml"
+String libraryName = "flutter-fire-ml-cus"
 File pubspec = new File(project.projectDir.parentFile,  'pubspec.yaml')
 
 if (pubspec.exists()) {
