+++ packages/firebase_ml/example/android/app/build.gradle
@@ -36,7 +36,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "io.flutter.plugins.firebaseml_example"
+        applicationId "io.flutter.plugins.firebasemlcustomexample"
         minSdkVersion 21
         targetSdkVersion 28
         versionCode flutterVersionCode.toInteger()
