+++ packages/firebase_ml/example/android/app/src/main/java/io/flutter/plugins/firebasemlcustomexample/EmbeddingV1Activity.java
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package io.flutter.plugins.firebaseml_example;
+package io.flutter.plugins.firebasemlcustomexample;
 
 import android.os.Bundle;
 import io.flutter.app.FlutterActivity;
