+++ packages/firebase_ml/example/android/app/src/main/java/io/flutter/plugins/firebasemlcustomexample/MainActivity.java
@@ -1,9 +1,9 @@
-package io.flutter.plugins.firebaseml_example;
+package io.flutter.plugins.firebasemlcustomexample;
 
 import io.flutter.embedding.android.FlutterActivity;
 import io.flutter.embedding.engine.FlutterEngine;
 import io.flutter.embedding.engine.plugins.shim.ShimPluginRegistry;
-import io.flutter.plugins.firebaseml.FirebaseMLPlugin;
+import io.flutter.plugins.firebasemlcustom.FirebaseMLCustomPlugin;
 import io.flutter.plugins.imagepicker.ImagePickerPlugin;
 import io.flutter.plugins.pathprovider.PathProviderPlugin;
 import sq.flutter.tflite.TflitePlugin;
@@ -11,7 +11,7 @@ import sq.flutter.tflite.TflitePlugin;
 public class MainActivity extends FlutterActivity {
   @Override
   public void configureFlutterEngine(FlutterEngine flutterEngine) {
-    flutterEngine.getPlugins().add(new FirebaseMLPlugin());
+    flutterEngine.getPlugins().add(new FirebaseMLCustomPlugin());
 
     final ShimPluginRegistry shimPluginRegistry = new ShimPluginRegistry(flutterEngine);
     PathProviderPlugin.registerWith(
