+++ packages/firebase_ml/example/test_driver/firebase_ml_custom.dart
@@ -6,7 +6,7 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 import 'package:flutter_driver/driver_extension.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:firebase_ml/firebase_ml.dart';
+import 'package:firebase_ml_custom/firebase_ml_custom.dart';
 import 'dart:io' show Platform;
 
 part 'firebase_model_manager.dart';
