+++ packages/firebase_ml/example/test_driver/firebase_model_manager.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of 'firebase_ml.dart';
+part of 'firebase_ml_custom.dart';
 
 void firebaseModelManagerTest() {
   group('$FirebaseModelManager', () {
