+++ packages/firebase_ml/lib/firebase_ml_custom.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-library firebase_ml;
+library firebase_ml_custom;
 
 export 'src/model_management/firebase_model_manager.dart';
 export 'src/model_management/firebase_model_download_conditions.dart';
